requirements.txt
asgiref==3.9.1
boto3==1.40.34
botocore==1.40.34
click==8.2.1
dj-database-url==3.0.1
Django==5.2.5
django-storages==1.14.6
git-filter-repo==2.47.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv  # <-- MAKE SURE THIS LINE IS HERE
razorpay
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svgwrite==1.4.3
Tree==0.2.4
urllib3==2.5.0
whitenoise==6.11.0